# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Pokedex backend",
    "description": "This is a pokedex api with OAuth2 login",
    "version": "1.0.0",
    "contact": {
      "name": "Theo Declerck",
      "email": "theo.dclrck5@gmail.com"
    }
  },
  "host": "localhost:8000",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/authorize": {
      "get": {
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/oauth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "client",
            "required": true,
            "schema": {
              "properties": {
                "login": {
                  "type": "string",
                  "example": "leopkmn"
                },
                "password": {
                  "type": "string",
                  "example": "cynthia"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "connected"
          },
          "400": {
            "description": "No authorization code provided"
          },
          "401": {
            "description": "Application is not authorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/exchange/create/self": {
      "post": {
        "tags": [
          "exchange"
        ],
        "summary": "Create an Exchange by 2 Trainer ID and 2 Pokemon ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Exchange",
            "required": true,
            "schema": {
              "properties": {
                "receiverId": {
                  "type": "integer"
                },
                "senderPokemonId": {
                  "type": "integer"
                },
                "receiverPokemonId": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Exchange created"
          },
          "400": {
            "description": "Exchange already exist"
          },
          "404": {
            "description": "Pokemon not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/exchange/create/{trainerId}": {
      "post": {
        "tags": [
          "exchange"
        ],
        "summary": "Create an Exchange by 2 Trainer ID and 2 Pokemon ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Exchange",
            "required": true,
            "schema": {
              "properties": {
                "receiverId": {
                  "type": "integer"
                },
                "senderPokemonId": {
                  "type": "integer"
                },
                "receiverPokemonId": {
                  "type": "integer"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "trainerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "description": "Exchange created"
          },
          "400": {
            "description": "Exchange already exist"
          },
          "404": {
            "description": "Pokemon not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/exchange/getSentExchanges/self": {
      "get": {
        "tags": [
          "exchange"
        ],
        "summary": "Returns all Exchange sent by the current logged in Trainer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Exchange"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/exchange/getSentExchanges/{trainerId}": {
      "get": {
        "tags": [
          "exchange"
        ],
        "summary": "Returns all Exchange sent by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "trainerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Exchange"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/exchange/getReceivedExchanges/self": {
      "get": {
        "tags": [
          "exchange"
        ],
        "summary": "Returns all Exchange received by the current logged in Trainer.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Exchange"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/exchange/getReceivedExchanges/{trainerId}": {
      "get": {
        "tags": [
          "exchange"
        ],
        "summary": "Returns all Exchange received by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "trainerId",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Exchange"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/exchange/sendResponse": {
      "patch": {
        "tags": [
          "exchange"
        ],
        "summary": "Update the status of an Exchange by ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Status",
            "required": true,
            "schema": {
              "properties": {
                "status": {
                  "properties": {
                    "accepted": {
                      "type": "string"
                    },
                    "refused": {
                      "type": "string"
                    }
                  }
                },
                "id": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/definitions/Exchange"
          },
          "400": {
            "description": "Exchange is no more available"
          },
          "404": {
            "description": "Exchange not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Exchange": {
      "properties": {
        "id": {
          "type": "integer"
        },
        "status": {
          "properties": {
            "waiting": {
              "type": "string"
            },
            "accepted": {
              "type": "string"
            },
            "refused": {
              "type": "string"
            }
          }
        },
        "senderId": {
          "type": "integer"
        },
        "receiverId": {
          "type": "integer"
        },
        "senderPokemonId": {
          "type": "integer"
        },
        "receiverPokemonId": {
          "type": "integer"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Token",
      "in": "header"
    }
  }
}